ddns-dynu/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
dnspython==2.2.1
idna==3.3
requests==2.27.1
urllib3==1.26.9
